# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Search, data quality, duplicate detection, and cross-table profile\nlookups over a 15M-row customer dataset (CHALLENGE.md Rounds 1-5).",
        "title": "17an Customer Intelligence API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/duplicates": {
            "post": {
                "description": "Top-level \"WAJIB\" endpoint: pairs of users sharing the same email or phone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "Exact duplicate pairs",
                "parameters": [
                    {
                        "description": "limit (default 100, max 1000)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "limit": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/duplicates/find": {
            "get": {
                "description": "Round 4: groups of users sharing the same IP address (ws_user_activity)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "Duplicate account clusters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "currently only ip_address",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 50, max 500",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/duplicates/{user_id}": {
            "get": {
                "description": "Submission-doc alias: other users sharing this user's email/phone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "Duplicate candidates for one user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "description": "Round 1: DB connectivity + cached total row count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/metrics": {
            "get": {
                "description": "Top-level \"WAJIB\" shape: duplicates, missing_fields, quality_score",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quality"
                ],
                "summary": "Slim quality metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/quality": {
            "get": {
                "description": "Round 3: missing/duplicate/invalid-format breakdown across email, phone, birth_date, hobbies, status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quality"
                ],
                "summary": "Data quality report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/search": {
            "get": {
                "description": "Round 2: exact match for email/phone/user_id, fuzzy trigram match for name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search customers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search term",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email | phone | user_id | name",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "default 10, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/user-profile/{user_id}": {
            "get": {
                "description": "Round 5: user + order count/total transaction amount + activity count/last activity (4-table JOIN)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Cross-table user profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 20, max 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get one user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}